Added functionality for some buttons.
Now, when a user interacts with the Watch, Like, Watchlist and Review buttons, the movie data will be stored in the database, and in the case of a review, the text written by the user and optionally a score will also be stored. All buttons will change their appearance depending on their state, for example, if a movie is liked, the color of the button will change to red and when the user hovers over it, the word dislike will appear, so when the user presses it, the movie will be removed from the database.